tests/wpt/tests/css/css-shapes/shape-functions/rect-function-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Shapes Module Level 1: parsing the rect() function</title>
<link rel="help" href="https://drafts.csswg.org/css-shapes-1/#funcdef-basic-shape-rect">
<meta name="assert" content="Tests parsing of the inset() function">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("clip-path", "rect(0 0 0 0)", "rect(0px 0px 0px 0px)");
test_valid_value("clip-path", "rect(10px auto 20px 30px)");
test_valid_value("clip-path", "rect(auto auto auto auto)");
test_valid_value("clip-path", "rect(10% 20% 15% 12%)");
test_valid_value("clip-path", "rect(10% 95% 97% 12%)");
test_valid_value("clip-path", "rect(-10% -20% -15% -12%)");
test_valid_value("clip-path", "rect(10px 12% 20px 30px round 100%)");
test_valid_value("clip-path", "rect(10px 12% 20px 30px round 0 1px)", "rect(10px 12% 20px 30px round 0px 1px)");
test_valid_value("clip-path", "rect(10px 12% 20px 30px round 0px 1px 2%)");
test_valid_value("clip-path", "rect(10px 12% 20px 30px round 0px 1px 2% 3em)");
test_valid_value("clip-path", "rect(10px 12% 20px 30px round 20% / 0px 1px 2% 3em)");
</script>
</body>
</html>
